+++ logbus/formatter/formatter.go
@@ -83,7 +83,7 @@ type Formatter struct {
 }
 
 // New creates a new Formatter.
-func New(ctx context.Context, b *logbus.Bus, debug, verbose, displayStats, forceColor, noColor, disableOngoingUpdates bool, execStatsTracker *execstatssummary.Tracker) *Formatter {
+func New(ctx context.Context, b *logbus.Bus, debug, verbose, displayStats, forceColor, noColor, disableOngoingUpdates bool, execStatsTracker *execstatssummary.Tracker, isGitHubActions bool) *Formatter {
 	ongoingTick := durationBetweenOngoingUpdatesNoAnsi
 	if ansiSupported {
 		ongoingTick = durationBetweenOngoingUpdates
@@ -110,7 +110,7 @@ func New(ctx context.Context, b *logbus.Bus, debug, verbose, displayStats, force
 	}
 	f := &Formatter{
 		bus:              b,
-		console:          conslogging.New(nil, nil, colorMode, conslogging.DefaultPadding, logLevel),
+		console:          conslogging.New(nil, nil, colorMode, conslogging.DefaultPadding, logLevel, isGitHubActions),
 		verbose:          verbose,
 		displayStats:     displayStats,
 		execStatsTracker: execStatsTracker,
@@ -247,6 +247,7 @@ func (f *Formatter) handleDeltaManifest(dm *logstream.DeltaManifest) error {
 	}
 	if dm.GetFields().GetHasFailure() {
 		f.printBuildFailure()
+		f.printGHAFailure()
 	}
 	return nil
 }
@@ -466,6 +467,54 @@ func (f *Formatter) printBuildFailure() {
 	f.lastOutputWasOngoingUpdate = false
 	f.lastOutputWasProgress = false
 	f.lastCommandOutput = nil
+}
+
+func (f *Formatter) printGHAFailure() {
+	failure := f.manifest.GetFailure()
+	if failure.GetErrorMessage() == "" {
+		return
+	}
+	var cm *logstream.CommandManifest
+	if failure.GetCommandId() != "" {
+		cm = f.manifest.GetCommands()[failure.GetCommandId()]
+	}
+	c, _ := f.targetConsole(failure.GetTargetId(), failure.GetCommandId(), false)
+	c = c.WithFailed(true)
+
+	// Extract the error from first line of the error message
+	lines := strings.Split(failure.GetErrorMessage(), "\n")
+	var message string
+	if len(lines) > 1 {
+		message = strings.Join(strings.Fields(strings.Join(lines[1:], " ")), " ")
+	} else {
+		message = strings.Join(strings.Fields(lines[0]), " ")
+	}
+
+	// Print GHA Error with line info if available
+	if cm != nil && cm.SourceLocation != nil &&
+		cm.SourceLocation.File != "" && cm.SourceLocation.StartLine > 0 {
+		c.PrintGHAError(message, conslogging.WithGHASourceLocation(cm.SourceLocation.File, cm.SourceLocation.StartLine, cm.SourceLocation.StartColumn))
+	} else {
+		c.PrintGHAError(message)
+	}
+
+	//GHA Summary markdown
+	markdown := fmt.Sprintf(`
+# ❌ Build Failure ❌
+
+### Error Message 
+
+~~~
+%s
+~~~
+
+### Command Output
+
+~~~
+%s
+~~~
+`, failure.GetErrorMessage(), string(failure.GetOutput()))
+	c.PrintGHASummary(markdown)
 }
 
 func (f *Formatter) targetName(targetID string) string {