+++ util/llbutil/authprovider/multiauthprovider_test.go
@@ -18,7 +18,7 @@ import (
 )
 
 func newConsLogger() conslogging.ConsoleLogger {
-	return conslogging.New(os.Stderr, &sync.Mutex{}, conslogging.NoColor, 0, conslogging.Info)
+	return conslogging.New(os.Stderr, &sync.Mutex{}, conslogging.NoColor, 0, conslogging.Info, false)
 }
 
 func TestMultiAuth(t *testing.T) {